First commit.  Started with LogsToDynamo which read some files from S3 bucket, parsed those files and sent data to a DynamoDB table.
This one removes the S3 and file parse functionality.  Instead this Lambda will be triggered by a CloudWatch event where certain strings are found. The CloudWatch entry will be passed straight to the Lambda as "event".  That event data is parsed and entered into the DynamoDB table.

Ready to upload and start playing with tomorrow...